Enable reading docker config from environment variable (#428)
When using `testcontainers` in a docker in docker environment it can be
a bit stainious to setup config files. GitLab alread suggests using a
environment variable DOCKER_AUTH_CONFIG to configure docker.

With this fix `testcontainers` checks if the DOCKER_AUTH_CONFIG is
available after trying all other means of authentication.